+++ Assets/Plugins/Mediabox/Source/API/IosMediaboxServer.cs
@@ -8,6 +8,7 @@ namespace Mediabox.API {
         [DllImport("__Internal")] static extern void OnLoadingFailed();
         [DllImport("__Internal")] static extern void OnUnloadingSucceeded();
         [DllImport("__Internal")] static extern void OnSaveDataWritten();
+        [DllImport("__Internal")] static extern void OnUserScoreChanged(float newValue);
         [DllImport("__Internal")] static extern void OnCreateScreenshotSucceeded(string path);
         [DllImport("__Internal")] static extern void OnCreateScreenshotFailed();
         [DllImport("__Internal")] static extern void OnGameExitRequested();
@@ -46,6 +47,12 @@ namespace Mediabox.API {
         void IMediaboxServer.OnCreateScreenshotSucceeded(string path) {
 #if UNITY_IOS
             OnCreateScreenshotSucceeded(path);
+#endif
+        }
+
+        public void OnUserScoreChanged(float newValue) {
+#if UNITY_IOS
+            OnUserScoreChanged(newValue);
 #endif
         }