+++ Assets/Plugins/Mediabox/Source/GameKit/GameManager/IGameAPI.cs
@@ -2,9 +2,9 @@ using System.Threading.Tasks;
 using UnityEngine.SceneManagement;
 
 namespace Mediabox.GameKit.GameManager {
-	public interface IGameManager {
-		void QuitApplication();
-		bool HasContentBundle { get; }
+	public interface IGameAPI {
+		void Quit();
+		void ReportNewUserScore(float newScore);
 		Task<T> LoadAssetFromContentBundle<T>(string assetPath) where T:UnityEngine.Object;
 		Task LoadSceneFromContentBundle<T>(string assetPath, LoadSceneMode loadSceneMode = LoadSceneMode.Single) where T:UnityEngine.Object;
 	}